Tridm

+++ src/main/java/com/swpproject/application/service/SlotExcerciseEntityService.java
@@ -36,7 +36,7 @@ public class SlotExcerciseEntityService {
                                             ){
         return slotExcerciseEntityRepository.findAllByMinWeekAndConditions(orderId,isPending);
     }
-    public void updateSlotOrderPending(Long slotExerciseId,
+    public void updateSlotOrderPending(Integer slotExerciseId,
                                        boolean isPending){
         slotExcerciseEntityRepository.updateIsPending(slotExerciseId,isPending);
     }
